Added favoriteFood slice state to Local Storage.
